+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add search functionality to allow users to filter restaurants by name, address or description
+
 ## [1.10.4] - 2024-09-28
 
 ### Fixed